current build

+++ Aeona Sprint 2/Aeona Project/source/AI_States/CRandomAIState.cpp
@@ -33,8 +33,8 @@ void CRandomAIState::DeleteInstance(void)
 
 void CRandomAIState::Enter(CBaseCharacter* pCharacter)
 {
-	pCharacter->SetVelY(1 * pCharacter->GetSpeed());
-	pCharacter->SetVelX(1 * pCharacter->GetSpeed());
+	pCharacter->SetVelY(1.0f * (float)pCharacter->GetSpeed());
+	pCharacter->SetVelX(1.0f * (float)pCharacter->GetSpeed());
 	pCharacter->SetCurrentAnimation(CBaseCharacter::ANM_WALK_RIGHT);
 }
